Initial Commit
This is the initial commit of the Intuitiontistic Theorem Prover. This
code was taken from a branch in my Modal Theorem Prover Repo which I
commandeered for a time to do this work. It's progressed far enough
that it deserves it's own repo and will get it. The code currently
does not compile because there are some unimplemented
functions. This commit is simply to get the base set of files that I
want in the repo recorded.