Add Tests
Still on Master, but now we have the test suite that we were using
before in Hspec. Plus, we have the failing test that began this
migration. I've made some small theoretical changes to Model.hs, but
nothing serious or plugged in yet. Soon it will be time to create a
develop branch and start fixing things.