Bugfix: Model Generation
This commit fixes the case where we were incoreectrly identifying
models as counterexamples. This caused a slow down to the point of not
finishing in some other test cases. A separate branch will be used to
address this since these test cases were always slow. Since the bug is
fixed, I'm committing this branch and will create a PR