Feature: Resolution Modules
A Resolution Module captures a pattern that is often seen when proving
things to prevent the prover from going down rabbit holes. For
example, we often saw the pattern G => (Necessarily p), (Possibly (Not
p)), S but because of formulas that were around that one, the prover
would not prune branches quickly enough to close the tree. The
addition of resolution modules allows the prover to look for specific
patterns and trim a proof tree based on them.

A resolution module has two parts that need to be specified:

   1.  The pattern that it is going to look for -- this is at the
   level of a sequent
   2.  The rules that it is going to apply to the hypersequent that
   it applies to to turn it into a closed branch.